+++ features/admin.users.v1/components/user-profile.tsx
@@ -18,8 +18,6 @@
 
 import Alert from "@oxygen-ui/react/Alert";
 import { Show, useRequiredScopes } from "@wso2is/access-control";
-import { getAllExternalClaims } from "@wso2is/admin.claims.v1/api";
-import { ClaimManagementConstants } from "@wso2is/admin.claims.v1/constants/claim-management-constants";
 import { AppConstants } from "@wso2is/admin.core.v1/constants/app-constants";
 import { history } from "@wso2is/admin.core.v1/helpers/history";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models/config";
@@ -40,7 +38,6 @@ import { getUserNameWithoutDomain, resolveUserstore } from "@wso2is/core/helpers
 import {
     AlertInterface,
     AlertLevels,
-    ExternalClaim,
     MultiValueAttributeInterface,
     ProfileInfoInterface,
     ProfileSchemaInterface,
@@ -220,8 +217,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
     const isCurrentUserAdmin: boolean = user?.roles?.some((role: RolesMemberInterface) =>
         role.display === administratorConfig.adminRoleName) ?? false;
 
-    const [ duplicatedUserClaims, setDuplicatedUserClaims ] = useState<ExternalClaim[]>([]);
-
     useEffect(() => {
         if (connectorProperties && Array.isArray(connectorProperties) && connectorProperties?.length > 0) {
 
@@ -278,54 +273,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
     useEffect(() => {
         mapUserToSchema(profileSchema, user);
     }, [ profileSchema, user ]);
-
-    /**
-     * This useEffect identifies external claims that are mapped to the same local claim
-     * between the Enterprise schema and the WSO2 System schema.
-     *
-     * These dual mappings occur only in migrated environments due to the SCIM2 schema restructuring
-     * introduced in https://github.com/wso2/product-is/issues/20850.
-     *
-     * The effect fetches both sets of external claims and detects overlaps by comparing their
-     * mappedLocalClaimURI values.
-     * Identified Enterprise claims are then excluded from the user profile UI to avoid redundancy.
-     */
-    useEffect(() => {
-        const fetchAllClaims = async () => {
-            try {
-                const [ enterpriseClaims, systemClaims ] = await Promise.all([
-                    getAllExternalClaims(
-                        ClaimManagementConstants.ATTRIBUTE_DIALECT_IDS.get("SCIM2_SCHEMAS_EXT_ENT_USER"),
-                        null
-                    ),
-                    getAllExternalClaims(
-                        ClaimManagementConstants.ATTRIBUTE_DIALECT_IDS.get("SCIM2_SCHEMAS_EXT_SYSTEM"),
-                        null
-                    )
-                ]);
-
-                const systemMappedClaimURIs: Set<string> = new Set(
-                    systemClaims.map((claim: ExternalClaim) => claim.mappedLocalClaimURI).filter(Boolean)
-                );
-                const duplicates: ExternalClaim[] = enterpriseClaims.filter(
-                    (claim: ExternalClaim) =>
-                        claim.mappedLocalClaimURI &&
-                        systemMappedClaimURIs.has(claim.mappedLocalClaimURI)
-                );
-
-                setDuplicatedUserClaims(duplicates);
-
-            } catch (error) {
-                dispatch(addAlert({
-                    description: t("claims:external.notifications.fetchExternalClaims.genericError.description"),
-                    level: AlertLevels.ERROR,
-                    message: t("claims:external.notifications.fetchExternalClaims.genericError.message")
-                }));
-            }
-        };
-
-        fetchAllClaims();
-    }, []);
 
     /**
      * The following function maps profile details to the SCIM schemas.
@@ -633,52 +580,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
         }
     };
 
-    // useEffect(() => {
-    //     mapMultiValuedAttributeValues(profileInfo);
-    // }, [ profileInfo ]);
-
-    // /**
-    //  * The following function map multi-valued attribute values and their primary values from profile data.
-    //  *
-    //  * @param profileData - Profile data.
-    //  */
-    // const mapMultiValuedAttributeValues = (profileData: Map<string, string>): void => {
-
-    //     const tempMultiValuedAttributeValues: Record<string, string[]> = {};
-    //     const tempPrimaryValues: Record<string, string> = {};
-
-    //     simpleMultiValuedExtendedProfileSchema?.forEach((schema: ProfileSchemaInterface) => {
-    //         const attributeValue: string = profileData?.get(schema.name);
-
-    //         tempMultiValuedAttributeValues[schema.name] = attributeValue ? attributeValue.split(",") : [];
-    //     });
-
-    //     if (isMultipleEmailAndMobileNumberEnabled) {
-    //         const primaryEmail: string = profileData?.get(EMAIL_ATTRIBUTE);
-    //         const primaryMobile: string = profileData?.get(MOBILE_ATTRIBUTE);
-
-    //         if (!isEmpty(primaryEmail)) {
-    //             const emailAddresses: string[] = tempMultiValuedAttributeValues[EMAIL_ADDRESSES_ATTRIBUTE]
-    //                 ?.filter((value: string) => !isEmpty(value) && value !== primaryEmail) ?? [];
-
-    //             emailAddresses.unshift(primaryEmail);
-    //             tempMultiValuedAttributeValues[EMAIL_ADDRESSES_ATTRIBUTE] = emailAddresses;
-    //         }
-    //         if (!isEmpty(primaryMobile)) {
-    //             const mobileNumbers: string[] = tempMultiValuedAttributeValues[MOBILE_NUMBERS_ATTRIBUTE]
-    //                 ?.filter((value: string) => !isEmpty(value) && value !== primaryMobile) ?? [];
-
-    //             mobileNumbers.unshift(primaryMobile);
-    //             tempMultiValuedAttributeValues[MOBILE_NUMBERS_ATTRIBUTE] = mobileNumbers;
-    //         }
-    //         tempPrimaryValues[EMAIL_ATTRIBUTE] = primaryEmail;
-    //         tempPrimaryValues[MOBILE_ATTRIBUTE] = primaryMobile;
-    //     }
-
-    //     setMultiValuedAttributeValues(tempMultiValuedAttributeValues);
-    //     setPrimaryValues(tempPrimaryValues);
-    // };
-
     /**
      * This function handles deletion of the user.
      *
@@ -746,44 +647,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
     };
 
     /**
-     * The function returns the normalized format of locale.
-     *
-     * @param locale - locale value.
-     * @param localeJoiningSymbol - symbol used to join language and region parts of locale.
-     * @param updateSupportedLanguage - If supported languages needs to be updated with the given localString or not.
-     */
-    /*const normalizeLocaleFormat = (
-        locale: string,
-        localeJoiningSymbol: LocaleJoiningSymbol,
-        updateSupportedLanguage: boolean
-    ): string => {
-        if (!locale) {
-            return locale;
-        }
-
-        const separatorIndex: number = locale.search(/[-_]/);
-
-        let normalizedLocale: string = locale;
-
-        if (separatorIndex !== -1) {
-            const language: string = locale.substring(0, separatorIndex).toLowerCase();
-            const region: string = locale.substring(separatorIndex + 1).toUpperCase();
-
-            normalizedLocale = `${language}${localeJoiningSymbol}${region}`;
-        }
-
-        if (updateSupportedLanguage && !supportedI18nLanguages[normalizedLocale]) {
-            supportedI18nLanguages[normalizedLocale] = {
-                code: normalizedLocale,
-                name: UserManagementConstants.GLOBE,
-                namespaces: []
-            };
-        }
-
-        return normalizedLocale;
-    };*/
-
-    /**
      * This function returns the ID of the administrator role.
      *
      */
@@ -820,30 +683,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
                 }));
             });
     };
-
-    /**
-     * This function returns the verification pending attribute value for email and mobile attributes.
-     * @param schemaName - Schema name.
-     * @returns Verification pending attribute value.
-     */
-    /*const getVerificationPendingAttributeValue = (schemaName: string): string | null => {
-        if (schemaName === EMAIL_ATTRIBUTE || schemaName === EMAIL_ADDRESSES_ATTRIBUTE) {
-            const pendingAttributes: Array<{value: string}> | undefined =
-            user[ProfileConstants.SCIM2_SYSTEM_USER_SCHEMA]?.["pendingEmails"];
-
-            return Array.isArray(pendingAttributes)
-                && pendingAttributes.length > 0
-                && pendingAttributes[0]?.value !== undefined
-                ? pendingAttributes[0].value
-                : null;
-        } else if (schemaName === MOBILE_ATTRIBUTE || schemaName === MOBILE_NUMBERS_ATTRIBUTE) {
-            return !isEmpty(user[ProfileConstants.SCIM2_SYSTEM_USER_SCHEMA]?.["pendingMobileNumber"])
-                ? user[ProfileConstants.SCIM2_SYSTEM_USER_SCHEMA]?.["pendingMobileNumber"]
-                : null;
-        }
-
-        return null;
-    };*/
 
     /**
      * This function handles deletion of the user.
@@ -895,522 +734,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
                 }));
             });
     };
-
-    /**
-     * Handles updating the primary email and mobile values when multiple emails and mobile numbers are enabled.
-     *
-     * @param values - The Map of form values.
-     */
-    /*const handlePrimaryEmailAndMobile = (values: Map<string, string | string[]>): void => {
-
-        const mobileSchema: ProfileSchemaInterface = profileSchema.find(
-            (schema: ProfileSchemaInterface) => schema.name === MOBILE_ATTRIBUTE
-        );
-
-        if (!isSchemaReadOnly(mobileSchema, isUserManagedByParentOrg)) {
-            const tempPrimaryMobile: string = primaryValues[MOBILE_ATTRIBUTE];
-            const mobileNumbersInputFieldValue: string = multiValuedInputFieldValue[MOBILE_NUMBERS_ATTRIBUTE];
-
-            if (tempPrimaryMobile !== undefined && tempPrimaryMobile !== null) {
-                values.set(MOBILE_ATTRIBUTE, tempPrimaryMobile);
-            }
-
-            if (isEmpty(tempPrimaryMobile) && !isEmpty(mobileNumbersInputFieldValue)) {
-                values.set(MOBILE_ATTRIBUTE, mobileNumbersInputFieldValue);
-            }
-        }
-
-        const emailSchema: ProfileSchemaInterface = profileSchema.find(
-            (schema: ProfileSchemaInterface) => schema.name === EMAIL_ATTRIBUTE
-        );
-
-        if (!isSchemaReadOnly(emailSchema, isUserManagedByParentOrg)) {
-            const tempPrimaryEmail: string = primaryValues[EMAIL_ATTRIBUTE] ?? "";
-            const emailsInputFieldValue: string = multiValuedInputFieldValue[EMAIL_ADDRESSES_ATTRIBUTE];
-
-            if (tempPrimaryEmail !== undefined && tempPrimaryEmail !== null) {
-                values.set(EMAIL_ATTRIBUTE, tempPrimaryEmail);
-            }
-            if (isEmpty(tempPrimaryEmail) && !isEmpty(emailsInputFieldValue)) {
-                values.set(EMAIL_ATTRIBUTE, emailsInputFieldValue);
-            }
-        }
-    };
-
-    const handleVerifiedEmailAddresses = (data: PatchRoleDataInterface): void => {
-
-        const verifiedAttributeSchema: ProfileSchemaInterface | undefined = profileSchema.find(
-            (schema: ProfileSchemaInterface) => schema.name === VERIFIED_EMAIL_ADDRESSES_ATTRIBUTE
-        );
-
-        if (!isMultipleEmailAndMobileNumberEnabled
-            || configSettings?.isEmailVerificationEnabled !== "true"
-            || isSchemaReadOnly(verifiedAttributeSchema, isUserManagedByParentOrg)) return;
-
-        const verifiedAttributeValueList: string[]
-            = profileInfo?.get(VERIFIED_EMAIL_ADDRESSES_ATTRIBUTE)?.split(",") ?? [];
-        const operation: ScimOperationsInterface = constructPatchOperationForMultiValuedVerifiedAttribute({
-            primaryValue: profileInfo?.get(EMAIL_ATTRIBUTE),
-            valueList: multiValuedAttributeValues[EMAIL_ADDRESSES_ATTRIBUTE],
-            verifiedAttributeSchema,
-            verifiedValueList: verifiedAttributeValueList
-        });
-
-        if (!operation) return;
-        data.Operations.push(operation);
-    };
-
-    const handleVerifiedMobileNumbers = (data: PatchRoleDataInterface): void => {
-        const verifiedAttributeSchema: ProfileSchemaInterface | undefined = profileSchema.find(
-            (schema: ProfileSchemaInterface) => schema.name === VERIFIED_MOBILE_NUMBERS_ATTRIBUTE
-        );
-
-        if (!isMultipleEmailAndMobileNumberEnabled
-            || configSettings?.isMobileVerificationEnabled !== "true"
-            || isSchemaReadOnly(verifiedAttributeSchema, isUserManagedByParentOrg)) return;
-
-        const verifiedAttributeValueList: string[]
-            = profileInfo?.get(VERIFIED_MOBILE_NUMBERS_ATTRIBUTE)?.split(",") ?? [];
-        const operation: ScimOperationsInterface = constructPatchOperationForMultiValuedVerifiedAttribute({
-            primaryValue: profileInfo?.get(MOBILE_ATTRIBUTE),
-            valueList: multiValuedAttributeValues[MOBILE_NUMBERS_ATTRIBUTE],
-            verifiedAttributeSchema,
-            verifiedValueList: verifiedAttributeValueList
-        });
-
-        if (!operation) return;
-        data.Operations.push(operation);
-    };
-    */
-
-    /**
-     * Removes the email address from the form values if it is pending verification.
-     * This prevents unnecessary retriggering of the verification process.
-     * @param values - Form values.
-     */
-    /*const handleVerificationPendingEmail = (values: Map<string, string | string[]>): void => {
-        if (isMultipleEmailAndMobileNumberEnabled
-            || configSettings?.isEmailVerificationEnabled !== "true"
-            || isEmpty(getVerificationPendingAttributeValue(EMAIL_ATTRIBUTE))
-            || values.get(EMAIL_ATTRIBUTE) !== getVerificationPendingAttributeValue(EMAIL_ATTRIBUTE)) return;
-
-        values.delete(EMAIL_ATTRIBUTE);
-    };*/
-
-    /**
-     * Removes the mobile number from the form values if it is pending verification.
-     * This prevents unnecessary retriggering of the verification process.
-     * @param values - Form values.
-     */
-    /*const handleVerificationPendingMobile = (values: Map<string, string | string[]>): void => {
-        if (isMultipleEmailAndMobileNumberEnabled
-            || configSettings?.isMobileVerificationEnabled !== "true"
-            || isEmpty(getVerificationPendingAttributeValue(MOBILE_ATTRIBUTE))
-            || values.get(MOBILE_ATTRIBUTE) !== getVerificationPendingAttributeValue(MOBILE_ATTRIBUTE)) return;
-
-        values.delete(MOBILE_ATTRIBUTE);
-    };*/
-
-    /**
-     * The following method handles the `onSubmit` event of forms.
-     *
-     * @param values - submit values.
-     */
-    /*const handleSubmit = (values: Map<string, string | string[]>): void => {
-
-        const data: PatchRoleDataInterface = {
-            Operations: [],
-            schemas: [ "urn:ietf:params:scim:api:messages:2.0:PatchOp" ]
-        };
-
-        let operation: ScimOperationsInterface = {
-            op: "replace",
-            value: {}
-        };
-
-        if (isMultipleEmailAndMobileNumberEnabled) {
-            handlePrimaryEmailAndMobile(values);
-            handleVerifiedEmailAddresses(data);
-            handleVerifiedMobileNumbers(data);
-        } else {
-            handleVerificationPendingEmail(values);
-            handleVerificationPendingMobile(values);
-        }
-
-        if (adminUserType === AdminAccountTypes.INTERNAL) {
-            profileSchema.forEach((schema: ProfileSchemaInterface) => {
-                const resolvedMutabilityValue: string = schema?.profiles?.console?.mutability ?? schema.mutability;
-                const sharedProfileValueResolvingMethod: string = schema?.sharedProfileValueResolvingMethod;
-
-                if (resolvedMutabilityValue === ProfileConstants.READONLY_SCHEMA || (isUserManagedByParentOrg &&
-                    sharedProfileValueResolvingMethod == SharedProfileValueResolvingMethod.FROM_ORIGIN)) {
-                    return;
-                }
-                if (!isFieldDisplayable(schema)) return;
-
-                let opValue: OperationValueInterface = {};
-
-                const schemaNames: string[] = schema.name.split(".");
-
-                if (schema.name !== "roles.default") {
-                    if (values.get(schema.name) !== undefined && values.get(schema.name).toString() !== undefined) {
-
-                        if (ProfileUtils.isMultiValuedSchemaAttribute(profileSchema, schemaNames[0]) ||
-                            schemaNames[0] === "phoneNumbers") {
-
-                            const attributeValues: (string | string[] | SchemaAttributeValueInterface)[] = [];
-                            const attValues: Map<string, string | string []> = new Map();
-
-                            if (schemaNames.length === 1 || schemaNames.length === 2) {
-                                if (schema.extended) {
-                                    opValue = {
-                                        [schema.schemaId]: constructPatchOpValueForMultiValuedAttribute(
-                                            schema.name,
-                                            multiValuedAttributeValues[schema.name],
-                                            multiValuedInputFieldValue[schema.name]
-                                        )
-                                    };
-                                } else {
-                                    // Handle emails and phoneNumbers and their sub attributes.
-                                    // Extract the sub attributes from the form values.
-                                    for (const value of values.keys()) {
-                                        const subAttribute: string[] = value.split(".");
-
-                                        if (subAttribute[0] === schemaNames[0]) {
-                                            attValues.set(value, values.get(value));
-                                        }
-                                    }
-
-                                    for (const [ key, value ] of attValues) {
-                                        const attribute: string[] = key.split(".");
-
-                                        if (value && value !== "") {
-                                            if (attribute.length === 1) {
-                                                attributeValues.push(value);
-                                            } else {
-                                                attributeValues.push({
-                                                    type: attribute[1],
-                                                    value: value
-                                                });
-                                            }
-                                        }
-                                    }
-                                    opValue = {
-                                        [schemaNames[0]]: attributeValues
-                                    };
-                                }
-                            }
-                        } else {
-                            if (schemaNames.length === 1) {
-                                if (schema.extended) {
-                                    const schemaId: string = schema?.schemaId
-                                        ? schema.schemaId
-                                        : userConfig.userProfileSchema;
-
-                                    if (schema.name === "externalId") {
-                                        opValue = {
-                                            [schemaNames[0]]: values.get(schemaNames[0])
-                                        };
-                                    } else {
-                                        opValue = {
-                                            [schemaId]: {
-                                                [schemaNames[0]]: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                    !!values.get(schema.name)?.includes(schema.name) :
-                                                    values.get(schemaNames[0])
-                                            }
-                                        };
-                                    }
-                                } else {
-                                    opValue = schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                        .get("EMAILS")
-                                        ? { emails: [ values.get(schema.name) ] }
-                                        : schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                            .get("LOCALE")
-                                            ? { [schemaNames[0]]: normalizeLocaleFormat(
-                                                values.get(schemaNames[0]) as string,
-                                                LocaleJoiningSymbol.UNDERSCORE,
-                                                false
-                                            ) }
-                                            : { [schemaNames[0]]: values.get(schemaNames[0]) };
-                                }
-                            } else {
-                                if (schema.extended && schema.multiValued) {
-                                    opValue = {
-                                        [schema.schemaId]: {
-                                            [schemaNames[0]]: constructPatchOpValueForMultiValuedAttribute(
-                                                schemaNames[1],
-                                                multiValuedAttributeValues[schema.name],
-                                                multiValuedInputFieldValue[schema.name]
-                                            )
-                                        }
-                                    };
-                                } else if (schema.extended) {
-                                    const schemaId: string = schema?.schemaId
-                                        ? schema.schemaId
-                                        : userConfig.userProfileSchema;
-
-                                    opValue = {
-                                        [schemaId]: {
-                                            [schemaNames[0]]: {
-                                                [schemaNames[1]]: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                    !!values.get(schema.name)?.includes(schema.name) :
-                                                    values.get(schema.name)
-                                            }
-                                        }
-                                    };
-                                } else if (schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                    .get("NAME")) {
-
-                                    if (values.get(schema.name) || values.get(schema.name) === "") {
-                                        opValue = {
-                                            name: { [schemaNames[1]]: values.get(schema.name) }
-                                        };
-                                    }
-                                } else {
-                                    if (schemaNames[0].includes("addresses")) {
-                                        if (schemaNames[0].split("#").length > 1) {
-                                            // Ex: addresses#home
-                                            const addressSchema: string = schemaNames[0]?.split("#")[0];
-                                            const addressType: string = schemaNames[0]?.split("#")[1];
-
-                                            opValue = {
-                                                [addressSchema]: [
-                                                    {
-                                                        type: addressType,
-                                                        [schemaNames[1]]: values.get(schema.name)
-                                                    }
-                                                ]
-                                            };
-                                        } else {
-                                            opValue = {
-                                                [schemaNames[0]]: [
-                                                    {
-                                                        formatted: values.get(schema.name),
-                                                        type: schemaNames[1]
-                                                    }
-                                                ]
-                                            };
-                                        }
-                                    } else if (schemaNames[0] !== "emails" && schemaNames[0] !== "phoneNumbers") {
-                                        opValue = {
-                                            [schemaNames[0]]: [
-                                                {
-                                                    type: schemaNames[1],
-                                                    value: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                        !!values.get(schema.name)?.includes(schema.name) :
-                                                        values.get(schema.name)
-                                                }
-                                            ]
-                                        };
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                operation = {
-                    op: "replace",
-                    value: opValue
-                };
-                // This is required as the api doesn't support patching the address attributes at the
-                // sub attribute level using 'replace' operation.
-                if (schemaNames[0].includes("addresses")) {
-                    operation.op = "add";
-                }
-                data.Operations.push(operation);
-            });
-
-        } else {
-            profileSchema.forEach((schema: ProfileSchemaInterface) => {
-                const resolvedMutabilityValue: string = schema?.profiles?.console?.mutability ?? schema.mutability;
-
-                if (resolvedMutabilityValue === ProfileConstants.READONLY_SCHEMA) {
-                    return;
-                }
-
-                if (!isFieldDisplayable(schema)) return;
-
-                let opValue: OperationValueInterface = {};
-
-                const schemaNames: string[] = schema.name.split(".");
-
-                if (schema.name !== "roles.default") {
-                    if (values.get(schema.name) !== undefined && values.get(schema.name).toString() !== undefined) {
-                        if (ProfileUtils.isMultiValuedSchemaAttribute(profileSchema, schemaNames[0]) ||
-                            schemaNames[0] === "phoneNumbers") {
-
-                            const attributeValues: (string | string[] | SchemaAttributeValueInterface)[] = [];
-                            const attValues: Map<string, string | string []> = new Map();
-
-                            if (schemaNames.length === 1 || schemaNames.length === 2) {
-
-                                if (schema.name === EMAIL_ADDRESSES_ATTRIBUTE
-                                    || schema.name == MOBILE_NUMBERS_ATTRIBUTE) {
-                                    opValue = {
-                                        [schema.schemaId]: constructPatchOpValueForMultiValuedAttribute(
-                                            schema.name,
-                                            multiValuedAttributeValues[schema.name],
-                                            multiValuedInputFieldValue[schema.name]
-                                        )
-                                    };
-                                } else {
-                                    // Extract the sub attributes from the form values.
-                                    for (const value of values.keys()) {
-                                        const subAttribute: string[] = value.split(".");
-
-                                        if (subAttribute[0] === schemaNames[0]) {
-                                            attValues.set(value, values.get(value));
-                                        }
-                                    }
-
-                                    for (const [ key, value ] of attValues) {
-                                        const attribute: string[] = key.split(".");
-
-                                        if (value && value !== "") {
-                                            if (attribute.length === 1) {
-                                                attributeValues.push(value);
-                                            } else {
-                                                attributeValues.push({
-                                                    type: attribute[1],
-                                                    value: value
-                                                });
-                                            }
-                                        }
-                                    }
-                                    opValue = {
-                                        [schemaNames[0]]: attributeValues
-                                    };
-                                }
-                            }
-                        } else {
-                            if (schemaNames.length === 1) {
-                                if (schema.extended) {
-                                    const schemaId: string = schema?.schemaId
-                                        ? schema.schemaId
-                                        : userConfig.userProfileSchema;
-
-                                    opValue = {
-                                        [schemaId]: {
-                                            [schemaNames[0]]: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                !!values.get(schema.name)?.includes(schema.name) :
-                                                values.get(schemaNames[0])
-                                        }
-                                    };
-                                } else {
-                                    opValue = schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                        .get("EMAILS")
-                                        ? { emails: [ values.get(schema.name) ] }
-                                        : schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                            .get("LOCALE")
-                                            ? { [schemaNames[0]]: normalizeLocaleFormat(
-                                                values.get(schemaNames[0]) as string,
-                                                LocaleJoiningSymbol.UNDERSCORE,
-                                                false
-                                            ) }
-                                            : { [schemaNames[0]]: values.get(schemaNames[0]) };
-                                }
-                            } else {
-                                if(schema.extended) {
-                                    const schemaId: string = schema?.schemaId
-                                        ? schema.schemaId
-                                        : userConfig.userProfileSchema;
-
-                                    opValue = {
-                                        [schemaId]: {
-                                            [schemaNames[0]]: {
-                                                [schemaNames[1]]: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                    !!values.get(schema.name)?.includes(schema.name) :
-                                                    values.get(schema.name)
-                                            }
-                                        }
-                                    };
-                                } else if (schemaNames[0] === UserManagementConstants.SCIM2_SCHEMA_DICTIONARY
-                                    .get("NAME")) {
-                                    opValue = {
-                                        name: { [schemaNames[1]]: values.get(schema.name) }
-                                    };
-                                } else {
-                                    if (schemaNames[0] === "addresses") {
-                                        opValue = {
-                                            [schemaNames[0]]: [
-                                                {
-                                                    formatted: values.get(schema.name),
-                                                    type: schemaNames[1]
-                                                }
-                                            ]
-                                        };
-                                    } else if (schemaNames[0] !== "emails" && schemaNames[0] !== "phoneNumbers") {
-                                        opValue = {
-                                            [schemaNames[0]]: [
-                                                {
-                                                    type: schemaNames[1],
-                                                    value: schema.type.toUpperCase() === "BOOLEAN" ?
-                                                        !!values.get(schema.name)?.includes(schema.name) :
-                                                        values.get(schema.name)
-                                                }
-                                            ]
-                                        };
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                operation = {
-                    op: "replace",
-                    value: opValue
-                };
-                // This is required as the api doesn't support patching the address attributes at the
-                // sub attribute level using 'replace' operation.
-                if (schemaNames[0] === "addresses") {
-                    operation.op = "add";
-                }
-                data.Operations.push(operation);
-            });
-        }
-
-        setIsSubmitting(true);
-
-        updateUserInfo(user.id, data)
-            .then(() => {
-                onAlertFired({
-                    description: t(
-                        "user:profile.notifications.updateProfileInfo.success.description"
-                    ),
-                    level: AlertLevels.SUCCESS,
-                    message: t(
-                        "user:profile.notifications.updateProfileInfo.success.message"
-                    )
-                });
-
-                handleUserUpdate(user.id);
-            })
-            .catch((error: AxiosError) => {
-                if (error?.response?.data?.detail || error?.response?.data?.description) {
-                    dispatch(addAlert({
-                        description: error?.response?.data?.detail || error?.response?.data?.description,
-                        level: AlertLevels.ERROR,
-                        message: t("user:profile.notifications.updateProfileInfo." +
-                            "error.message")
-                    }));
-
-                    return;
-                }
-
-                dispatch(addAlert({
-                    description: t("user:profile.notifications.updateProfileInfo." +
-                        "genericError.description"),
-                    level: AlertLevels.ERROR,
-                    message: t("user:profile.notifications.updateProfileInfo." +
-                        "genericError.message")
-                }));
-            })
-            .finally(() => {
-                setIsSubmitting(false);
-            });
-    };*/
 
     /**
      * Handle danger zone toggle actions.
@@ -1993,7 +1316,6 @@ export const UserProfile: FunctionComponent<UserProfilePropsInterface> = (
                         setIsUpdating={ (isUpdating: boolean): void => setIsSubmitting(isUpdating) }
                         onUserUpdate={ handleUserUpdate }
                         accountConfigSettings={ configSettings }
-                        duplicatedUserClaims={ duplicatedUserClaims }
                         data-componentid={ testId }
                     />
                 </EmphasizedSegment>
