+++ features/admin.users.v1/components/user-profile/legacy-user-profile-form.tsx
@@ -25,6 +25,7 @@ import Table from "@oxygen-ui/react/Table";
 import TableBody from "@oxygen-ui/react/TableBody";
 import TableCell from "@oxygen-ui/react/TableCell";
 import TableRow from "@oxygen-ui/react/TableRow";
+import { getAllExternalClaims } from "@wso2is/admin.claims.v1/api/claims";
 import { ClaimManagementConstants } from "@wso2is/admin.claims.v1/constants/claim-management-constants";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models/config";
 import { AppState } from "@wso2is/admin.core.v1/store";
@@ -100,7 +101,6 @@ interface UserProfileFormPropsInterface extends IdentifiableComponentInterface {
     isUserManagedByParentOrg?: boolean;
     setIsUpdating: (isUpdating: boolean) => void;
     accountConfigSettings: AccountConfigSettingsInterface;
-    duplicatedUserClaims: ExternalClaim[];
     onUpdate: (userId: string) => void;
 }
 
@@ -116,7 +116,6 @@ const UserProfileForm: FunctionComponent<UserProfileFormPropsInterface> = (
         setIsUpdating,
         onUserUpdate,
         accountConfigSettings,
-        duplicatedUserClaims,
         onUpdate,
         ["data-componentid"]: componentId = "user-profile-form"
     }: UserProfileFormPropsInterface
@@ -142,6 +141,7 @@ const UserProfileForm: FunctionComponent<UserProfileFormPropsInterface> = (
             useState<Record<string, string[]>>({});
     const [ multiValuedInputFieldValue, setMultiValuedInputFieldValue ] = useState<Record<string, string>>({});
     const [ primaryValues, setPrimaryValues ] = useState<Record<string, string>>({});
+    const [ duplicatedUserClaims, setDuplicatedUserClaims ] = useState<ExternalClaim[]>([]);
 
     const isDistinctAttributeProfilesFeatureEnabled: boolean = isFeatureEnabled(featureConfig?.attributeDialects,
         ClaimManagementConstants.DISTINCT_ATTRIBUTE_PROFILES_FEATURE_FLAG);
@@ -166,6 +166,57 @@ const UserProfileForm: FunctionComponent<UserProfileFormPropsInterface> = (
         return CommonUtils.getCountryList();
     }, []);
 
+    /**
+     * This useEffect identifies external claims that are mapped to the same local claim
+     * between the Enterprise schema and the WSO2 System schema.
+     *
+     * These dual mappings occur only in migrated environments due to the SCIM2 schema restructuring
+     * introduced in https://github.com/wso2/product-is/issues/20850.
+     *
+     * The effect fetches both sets of external claims and detects overlaps by comparing their
+     * mappedLocalClaimURI values.
+     * Identified Enterprise claims are then excluded from the user profile UI to avoid redundancy.
+     */
+    useEffect(() => {
+        const fetchAllClaims = async () => {
+            try {
+                const [ enterpriseClaims, systemClaims ] = await Promise.all([
+                    getAllExternalClaims(
+                        ClaimManagementConstants.ATTRIBUTE_DIALECT_IDS.get("SCIM2_SCHEMAS_EXT_ENT_USER"),
+                        null
+                    ),
+                    getAllExternalClaims(
+                        ClaimManagementConstants.ATTRIBUTE_DIALECT_IDS.get("SCIM2_SCHEMAS_EXT_SYSTEM"),
+                        null
+                    )
+                ]);
+
+                const systemMappedClaimURIs: Set<string> = new Set(
+                    systemClaims.map((claim: ExternalClaim) => claim.mappedLocalClaimURI).filter(Boolean)
+                );
+                const duplicates: ExternalClaim[] = enterpriseClaims.filter(
+                    (claim: ExternalClaim) =>
+                        claim.mappedLocalClaimURI &&
+                        systemMappedClaimURIs.has(claim.mappedLocalClaimURI)
+                );
+
+                setDuplicatedUserClaims(duplicates);
+
+            } catch (error) {
+                dispatch(addAlert({
+                    description: t("claims:external.notifications.fetchExternalClaims.genericError.description"),
+                    level: AlertLevels.ERROR,
+                    message: t("claims:external.notifications.fetchExternalClaims.genericError.message")
+                }));
+            }
+        };
+
+        fetchAllClaims();
+    }, []);
+
+    /**
+     * Set multi-valued attribute values to the state.
+     */
     useEffect(() => {
         mapMultiValuedAttributeValues(flattenedProfileData);
     }, [ flattenedProfileData, simpleMultiValuedExtendedProfileSchema ]);
